Shifted contents of __init__.py to app.py
